requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fonttools==4.39.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.21.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
pywin32==306
pyzmq==25.0.2
requests==2.28.2
six==1.16.0
stack-data==0.6.2
#If using CUDA, your version will be specific to your GPU
#-f https://download.pytorch.org/whl/torch_stable.html
#torch==1.13.1+cu116
torchmetrics==0.11.4
torchsummary==1.5.1
#If using CUDA, your version will be specific to your GPU
#-f https://download.pytorch.org/whl/torch_stable.html
#torchvision==0.14.1+cu116
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
